+++ app/services/analytics.ts
@@ -156,6 +156,7 @@ export default class Analytics extends Service {
         this: Analytics,
         pagePublic: boolean | undefined,
         resourceType: string,
+        versionType: string,
     ) {
         // Wait until everything has settled
         yield waitForQueue('destroy');
@@ -168,6 +169,7 @@ export default class Analytics extends Service {
         logEvent(this, 'Tracked page', {
             pagePublic,
             resourceType,
+            versionType,
             ...eventParams,
         });
 
@@ -177,6 +179,7 @@ export default class Analytics extends Service {
                 authenticated,
                 isPublic,
                 resource,
+                version,
             } = gaConfig.dimensions!;
 
             let isPublicValue = 'n/a';
@@ -196,6 +199,7 @@ export default class Analytics extends Service {
                 [authenticated]: this.session.isAuthenticated ? 'Logged in' : 'Logged out',
                 [isPublic]: isPublicValue,
                 [resource]: resourceType,
+                [version]: versionType,
                 ...eventParams,
             });
         }
@@ -243,8 +247,9 @@ export default class Analytics extends Service {
         this: Analytics,
         pagePublic?: boolean,
         resourceType: string = 'n/a',
+        version: string = 'n/a',
     ) {
-        this.get('trackPageTask').perform(pagePublic, resourceType);
+        this.get('trackPageTask').perform(pagePublic, resourceType, version);
     }
 
     trackFromElement(target: Element, initialInfo: InitialEventInfo) {