+++ lib/osf-components/addon/components/osf-dialog/trigger/component.ts
@@ -1,0 +1,10 @@
+import { tagName } from '@ember-decorators/component';
+import Component from '@ember/component';
+import hbs from 'htmlbars-inline-precompile';
+
+import { layout } from 'ember-osf-web/decorators/component';
+
+@tagName('')
+@layout(hbs`{{yield}}`)
+export default class OsfDialogTrigger extends Component {
+}