+++ mirage/serializers/taxonomy.ts
@@ -1,0 +1,6 @@
+// import { ModelInstance } from 'ember-cli-mirage';
+import Taxonomy from 'ember-osf-web/models/taxonomy';
+import ApplicationSerializer from './application';
+
+export default class TaxonomySerializer extends ApplicationSerializer<Taxonomy> {
+}